+++ src/config/settings/base.py
@@ -61,6 +61,7 @@ INSTALLED_APPS = [
     "axes",
     "csp",
     "djstripe",
+    "anymail",
     # Wagtail (optional)
 ] + WAGTAIL_APPS + [
     # First-party
@@ -138,7 +139,22 @@ AUTHENTICATION_BACKENDS = [
     "guardian.backends.ObjectPermissionBackend",
 ]
 
-SITE_ID = 1
+
+# Email Configuration
+EMAIL_BACKEND = env.str("EMAIL_BACKEND", default="django.core.mail.backends.smtp.EmailBackend")
+EMAIL_HOST = env.str("EMAIL_HOST", default="mailpit")
+EMAIL_PORT = env.int("EMAIL_PORT", default=1025)
+EMAIL_HOST_USER = env.str("EMAIL_HOST_USER", default="")
+EMAIL_HOST_PASSWORD = env.str("EMAIL_HOST_PASSWORD", default="")
+EMAIL_USE_TLS = env.bool("EMAIL_USE_TLS", default=False)
+DEFAULT_FROM_EMAIL = env.str("DEFAULT_FROM_EMAIL", default="noreply@acme.dev")
+SERVER_EMAIL = env.str("SERVER_EMAIL", default="server@acme.dev")
+
+ANYMAIL = {
+    "SENDGRID_API_KEY": env.str("ANYMAIL_API_KEY", default=""),
+    "WEBHOOK_SECRET": env.str("ANYMAIL_WEBHOOK_SECRET", default=""),
+}
+
 ACCOUNT_EMAIL_VERIFICATION = "optional"
 ACCOUNT_LOGIN_METHODS = {"email"}
 ACCOUNT_SIGNUP_FIELDS = ["email*", "password1*", "password2*"]